fixed bug with SmartEnemy in WaveGame - changed many ints to floats

+++ WaveGame/src/main/FastEnemy.java
@@ -30,11 +30,11 @@ public class FastEnemy extends GameObject {
 
 	public void render(Graphics g) {
 		g.setColor(Color.MAGENTA);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int) x, (int) y, 16, 16);
 	}
 
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int) x, (int) y, 16, 16);
 	}
 
 }
